requirements.txt
setuptools~=50.3.0
PyYAML~=5.3.1
docopt==0.6.2
